requirements.txt
streamlit>=1.29.0
google-generativeai>=0.4.0
PyMuPDF>=1.23.8
python-pptx>=0.6.21
pandas>=2.1.0
plotly>=5.17.0
